Specifying destination file for runURL, runGist, and runGitHub

+++ R/run-url.R
@@ -14,6 +14,9 @@
 #' @param subdir A subdirectory in the repository that contains the app. By
 #'   default, this function will run an app from the top level of the repo, but
 #'   you can use a path such as `\code{"inst/shinyapp"}.
+#' @param destdir Directory to store the downloaded application files. If \code{NULL}
+#'   (the default), the application files will be stored in a temporary directory
+#'   and removed when the app exits
 #' @param ... Other arguments to be passed to \code{\link{runApp}()}, such as
 #'   \code{port} and \code{launch.browser}.
 #' @export
@@ -26,7 +29,7 @@
 #'   runUrl("https://github.com/rstudio/shiny_example/archive/master.zip",
 #'     subdir = "inst/shinyapp/")
 #' }
-runUrl <- function(url, filetype = NULL, subdir = NULL, ...) {
+runUrl <- function(url, filetype = NULL, subdir = NULL, destdir = NULL, ...) {
 
   if (!is.null(subdir) && ".." %in% strsplit(subdir, '/')[[1]])
     stop("'..' not allowed in subdir")
@@ -44,8 +47,14 @@ runUrl <- function(url, filetype = NULL, subdir = NULL, ...) {
     stop("Unknown file extension.")
 
   message("Downloading ", url)
-  filePath <- tempfile('shinyapp', fileext=fileext)
-  fileDir  <- tempfile('shinyapp')
+  if (is.null(destdir)) {
+    filePath <- tempfile('shinyapp', fileext = fileext)
+    fileDir  <- tempfile('shinyapp')
+  } else {
+    fileDir <- destdir
+    filePath <- paste(destdir, fileext)
+  }
+
   dir.create(fileDir, showWarnings = FALSE)
   if (download(url, filePath, mode = "wb", quiet = TRUE) != 0)
     stop("Failed to download URL ", url)
@@ -65,8 +74,11 @@ runUrl <- function(url, filetype = NULL, subdir = NULL, ...) {
     first <- as.character(unzip(filePath, list=TRUE)$Name)[1]
     unzip(filePath, exdir = fileDir)
   }
-  on.exit(unlink(fileDir, recursive = TRUE), add = TRUE)
 
+  if(is.null(destdir)){
+    on.exit(unlink(fileDir, recursive = TRUE), add = TRUE)
+  }
+  
   appdir <- file.path(fileDir, first)
   if (!file_test('-d', appdir)) appdir <- dirname(appdir)
 
@@ -79,6 +91,9 @@ runUrl <- function(url, filetype = NULL, subdir = NULL, ...) {
 #'   https://gist.github.com/jcheng5/3239667, then \code{3239667},
 #'   \code{'3239667'}, and \code{'https://gist.github.com/jcheng5/3239667'} are
 #'   all valid values.
+#' @param destdir Directory to store the downloaded gist. If \code{NULL}
+#'   (the default), the application files will be stored in a temporary directory
+#'   and removed when the app exits
 #' @export
 #' @examples
 #' ## Only run this example in interactive R sessions
@@ -90,7 +105,7 @@ runUrl <- function(url, filetype = NULL, subdir = NULL, ...) {
 #'   runGist("https://gist.github.com/3239667")
 #' }
 #'
-runGist <- function(gist, ...) {
+runGist <- function(gist, destdir = NULL, ...) {
 
   gistUrl <- if (is.numeric(gist) || grepl('^[0-9a-f]+$', gist)) {
     sprintf('https://gist.github.com/%s/download', gist)
@@ -100,7 +115,7 @@ runGist <- function(gist, ...) {
     stop('Unrecognized gist identifier format')
   }
 
-  runUrl(gistUrl, filetype=".tar.gz", ...)
+  runUrl(gistUrl, filetype = ".tar.gz", destdir = destdir, ...)
 }
 
 
@@ -110,6 +125,9 @@ runGist <- function(gist, ...) {
 #'   \code{"username/repo"}, \code{username} will be taken from \code{repo}.
 #' @param ref Desired git reference. Could be a commit, tag, or branch name.
 #'   Defaults to \code{"master"}.
+#' @param destdir Directory to store the downloaded repository. If \code{NULL}
+#'   (the default), the application files will be stored in a temporary directory
+#'   and removed when the app exits
 #' @export
 #' @examples
 #' ## Only run this example in interactive R sessions
@@ -121,7 +139,7 @@ runGist <- function(gist, ...) {
 #'   runGitHub("shiny_example", "rstudio", subdir = "inst/shinyapp/")
 #' }
 runGitHub <- function(repo, username = getOption("github.user"),
-                      ref = "master", subdir = NULL, ...) {
+                      ref = "master", subdir = NULL, destdir = NULL, ...) {
 
   if (grepl('/', repo)) {
     res <- strsplit(repo, '/')[[1]]
@@ -133,5 +151,5 @@ runGitHub <- function(repo, username = getOption("github.user"),
   url <- paste("https://github.com/", username, "/", repo, "/archive/",
                ref, ".tar.gz", sep = "")
 
-  runUrl(url, subdir=subdir, ...)
+  runUrl(url, subdir = subdir, destdir = destdir, ...)
 }
